+++ lib/stripe/object_types.rb
@@ -9,6 +9,7 @@ module Stripe
       {
         # data structures
         ListObject::OBJECT_NAME => ListObject,
+        SearchResultObject::OBJECT_NAME => SearchResultObject,
 
         # business objects
         Account::OBJECT_NAME => Account,
@@ -94,6 +95,7 @@ module Stripe
         Terminal::ConnectionToken::OBJECT_NAME => Terminal::ConnectionToken,
         Terminal::Location::OBJECT_NAME => Terminal::Location,
         Terminal::Reader::OBJECT_NAME => Terminal::Reader,
+        TestHelpers::TestClock::OBJECT_NAME => TestHelpers::TestClock,
         ThreeDSecure::OBJECT_NAME => ThreeDSecure,
         Token::OBJECT_NAME => Token,
         Topup::OBJECT_NAME => Topup,