+++ lib/stripe/resources/payment_intent.rb
@@ -12,6 +12,7 @@ module Stripe
     custom_method :cancel, http_verb: :post
     custom_method :capture, http_verb: :post
     custom_method :confirm, http_verb: :post
+    custom_method :verify_microdeposits, http_verb: :post
 
     def cancel(params = {}, opts = {})
       request_stripe_object(
@@ -35,6 +36,15 @@ module Stripe
       request_stripe_object(
         method: :post,
         path: resource_url + "/confirm",
+        params: params,
+        opts: opts
+      )
+    end
+
+    def verify_microdeposits(params = {}, opts = {})
+      request_stripe_object(
+        method: :post,
+        path: resource_url + "/verify_microdeposits",
         params: params,
         opts: opts
       )